Added support for PySide2
As part of the VFX platform, Maya needs to upgrade its PySide component to
PySide2. These changes will make it so PyMEL uses either PySide or PySide2,
depending on what is available at runtime. In other words, the code works
now with PySide and will work with PySide2 when Maya starts shipping with
it.