stubs: trigger import of a module if the parent module of a class's base class is in maybe_modules
ie, in flux.ae.template, had:

import pymel.core as pm
class Template(pm.uitypes.AETemplate):
    ...

the "import pymel.core as pm" added an entry to maybe_modules... then when looking up the classname for
pm.uitypes.AETemplate, pymel.core.uitypes / pm.uitypes wasn't in maybe_modules (though pm was) - and
the import statement didn't get triggered

now we check if the parent module (in this case, pymel.core / pm) is in maybe_modules, and use it's import
text if so.

Note that (to be safe) we still need to add a "import pymel.core.uitypes" line, to ensure that ".uitypes" exists
on pymel.core (though in this particular case, this isn't necessary, in general, it might be)

this commits gets all stubs working in maya 2018!