nodetypes: import uitypes, for backward compatibility
there was a bug in the way that LazyLoadModule was used:
because data descriptors require items to be installed
at the class level, and an instance of LazyLoadModule
was used directly in the pymel.core.uitypes module, this
meant that all it's lazy members were installed directly
on the LazyLoadModule base class.  Thus,
pymel.core.nodetypes inherited them (even though IT used
a subclass, NodetypesLazyLoadModule, so IT'S members
weren't visible to uitypes).
The fix would have been to enforce creation of
subclasses, probably throuhg LazyLoadModule.__new__;
however, as the class isn't used anymore, not bothering
with this fix.

Debated not including these members, but there
is a decent chance someone was using them, even if on
accident, so we decided to include them