nodeclass.py: api methods with no args are never undoable
...because the api undo system is based around using a corresponding
"getter" to get the args, and then setting them with "this" function.
Hoewver, with no args, there is no way to set it to the results
returned by the getter.

Technically, an api function with no args could be it's own do and undo,
if it was a "toggle" type function... however, we currently have no way of
detecting that. So we assume a func with no args is not undoable.

This has no change in behavior, since no functions that were actually
wrapped were detected as "undoable" if they had no args.