build: put Enums behind conditionals based on maya version
...only if they're actually different in different versions

Also lays some groundwork if we wish to expand this other things besides
enums.  Considered trying to make it fully generic (ie, any attr or
method would get a conditional if it changed in maya versions), but that
seemed like too much work - particularly since we call factories.*
functions frequently, which assume there is a "singular" cache - in
theory, we'd need to find some way to use different cache versions (or
at least confirm that everywhere we do this, it's ok to use the
current version).