enable getRotateOrientation / setRotateOrientation
these are distinct from the existing getRotateAxis / setRotateAxis
(from MEL), because they operate on quaternions, which are what the
transform "actually" stores under the hood - and so it seems potentially
useful to give direct access.

Also, adding a manual wrap resolves a couple of messy issues -
formerly, getRotateOrientation was set to disable, and 'useName' of
_getRotateOrientation, which had a couple of problems - subclasses
still had getRotateOrientation / setRotateOrientation, but not only
that, the setRotateOrientation had it's pair set to
_getRotateOrientation, which didn't exist! Also, we're eliminated the
unused "space" argument.