LazyLoadModule - now automatically subclasses
this prevents pollution of the LazyLoadModule class's namespace, which
would be shared by all LazyLoadModules.  This is what happened with
pymel.core.uitypes - it installed it's lazy-load attrs on the base class,
LazyLoadModule.  pymel.core.nodetypes used it's own subclass -
NodetypesLazyLoadModule - but since that inherited from LazyLoadModule, the
net effect was that nodetypes had all the uitypes accessible from it's
namespace.

Since this bug has been out there for a while, I'm manually installing all
these uitypes in the nodetypes namespace to preserve backwards
compatibility... but we may try to remove these at some point.