some more str/unicode tweaks
in general, the desire is to keep behavior in python 2 similar to before
(ie, keep results `str` if they were `str`, unicode if they were unicode),
while balancing the desire to keep code 2+3 compatible, without making it
too ugly.

This means no uniformly using `unicode_literals`, or always importing
`builtins.str`, which would have the effect of making many things that were
formerly `str` be effectively `unicode`, which could lead to errors for
consumers.

For situations which formerly used `str`, leaving it using `str` will both keep
behavior the same in python-2, AND should "just work" (for the most part) in
python-3, so we generally aim to "leave str alone" - ie, no conversion to
unicode or `builtins.str` in python-3.

What about things that were unicode?  Changing these to use python-2 `str` is
dangerous - for instance, any attempt to cast a `unicode` with non-ascii
characters to `str` will error.  So, we either need to make these stay "true"
`unicode`, or convert to `builtins.str`.

However, even a switch from `unicode` to `builtins.str` is potentially risky -
though they are both "effectively" unicode, as both will support the same
methods, and will both return True for `isinstance(x, unicode)`, they
differ in two ways:

- if someone does an explicit type check - ie, `type(x) is unicode` or
  `type(x) == unicode` - this will fail
- repr for `builtins.str` will not have a leading u - ie, `'foo'` not `u'foo'`

The first is something that generally "shouldn't" be done (but undoubtedly
is in many instances), while the second could be a factor if consumers are
using doctests.

However, using `builtins.str` has advantages in keeping code "cleanly"
six-compatible - ie, fewer ugly `if PY2` checks.  Therefore, I've tried to
strike a balance between usage of "true" python-2 `unicode` and `builtins.str`:

- In situations where something is going to be "widely used" - for
  instance, the ProxyUnicode method which DependNode uses - I keep it
  "true" `unicode`, for maximum compatibility
- If something already provides explicit str-vs-unicode branching behavior,
  I keep it using "true" ``unicode, on the premise that this may matter.
  For instance, classes which provide distinct `__str__` and `__unicode__`
  methods, I keep "true" `unicode`, and use `if PY2` branching.
- For other situations, I generally opt for `builtins.str`, for easier
  porting and readability.  In particular, if using unicode would require a lot
  of ugly if-branching, and it's not a really core method, I generally switch to
  `builtins.str`.

For convience, in cases where `builtins.str` is used a lot in a module (and
pure `str` is not), I do a module-level `from builtins import str`.  However,
this does mean that what the `str` name represents varies a bit from module to
module.  Long term, this would be unacceptable, but since I view the py2/3
compatibility issues a temporary situation, I'm going to say this is ok for now.