Fix dir() for PyNode subclasses in Python 3
Python 3 has a __dir__ on str, which means any class that subclasses it will show
wrong results for dir().  Work around this by removing str.__dir__ from ProxyUnicode
so it uses the normal default behavior.

Before, the output of print(dir(pm.PyNode('persp'))) was the same as dir('').  This
brings it back to what it was in Python 2.